move mock to trait
